adds NSPhotoLibraryUsageDescription to info.plist, refactors some constraints in the editProfilevc
