+++ quavi/quavi/Controllers/Profile/Edit ProfileVC/EditProfileViewController.swift
@@ -198,42 +198,6 @@ class EditProfileViewController: UIViewController {
         emailTextField.delegate = self
     }
     
-    func setUpSubviews() {
-        self.view.addSubview(topBarView)
-        self.topBarView.addSubview(backButton)
-        self.topBarView.addSubview(confirmEditButton)
-        self.topBarView.addSubview(cancelEditButton)
-        self.view.addSubview(logoutButton)
-        
-        self.view.addSubview(userImage)
-        self.view.addSubview(changeImageButton)
-        self.view.addSubview(nameTextField)
-        self.view.addSubview(usernameTextField)
-        self.view.addSubview(emailTextField)
-        self.view.addSubview(nameLabel)
-        self.view.addSubview(usernameLabel)
-        self.view.addSubview(emailLabel)
-        self.view.addSubview(logoutButton)
-    }
-    
-    private func setUpConstraints() {
-        constrainTopBarView()
-        constrainBackButton()
-        constrainConfirmButton()
-        constrainCancelEditButton()
-        constrainLogoutButton()
-        
-        constrainUserImageView()
-        constrainChangeImageButton()
-        constrainTextFields(textField: nameTextField, textFieldAbove: nil)
-        constrainTextFields(textField: usernameTextField, textFieldAbove: nameTextField)
-        constrainTextFields(textField: emailTextField, textFieldAbove: usernameTextField)
-        constrainTextFieldLabels(label: nameLabel, textFieldBelow: nameTextField)
-        constrainTextFieldLabels(label: usernameLabel, textFieldBelow: usernameTextField)
-        constrainTextFieldLabels(label: emailLabel, textFieldBelow: emailTextField)
-        
-    }
-    
     private func setUpStyling() {
         styleTextViews(textfield: nameTextField)
         styleTextViews(textfield: usernameTextField)
