+++ quavi/quavi/Info.plist
@@ -63,5 +63,7 @@
 		<string>UIInterfaceOrientationLandscapeLeft</string>
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
+	<key>NSPhotoLibraryUsageDescription</key>
+	<string>This app uses selected photos from your photo library for your user image.</string>
 </dict>
 </plist>
